Removed travis.yml
Yeah, integration testing does not work. It's not about my code or layout,
it's about the usage of django-admin's templating option. So, with pushing
this repository, it will use the code inside of the repo. But you have to
use the repo as template to create an actual project. Then all tests in
tox.ini will work as expected.