Minimal adjustments in tox.ini
This branch is stalled right now... I don't really believe, that using tox for
testing is necessary in a Django project environment, because the versions of
Python and Django are fixed from the beginning and should not be object to
regular change.
The intra-project testing has to be performed against just one Python/Django
combination.

This is - obviously - different while developing/testing a single app, but
there are other ways to do just that.